Updating assignments/geometry/Geometry.java

+++ assignments/geometry/Geometry.java
@@ -32,7 +32,7 @@ public class Geometry {
     return Math.PI * r * Math.sqrt(Math.pow(r, 2) + Math.pow(h, 2)) + Math.PI * Math.pow(r, 2);
   }
   public double volumeOfCone (double r, double h){
-    return Math.PI * Math.pow(r, 2) * 1.0/3 * h;
+    return Math.PI * Math.pow(r, 2) * (double) 1/3 * h;
   }
   public double perimeterOfSquare(double side) {
     return side * 4;
